perf: never list-watch with cache penetration even on 410 lists
in case of apiserver cache avalanche, kelemetry is unnecessarily
requesting the apiserver to penetrate the cache for a full view to all
api objects, which further worsens the avalanche condition. meanwhile,
kelemetry itself would have preferred the earliest resource version
available on the apiserver, and penetrating the watch cache does not
help.